LayoutTests/css3/flexbox/flexbox-ignore-container-firstLetter.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
    .container { display: -webkit-flex }
    .container::first-letter { line-height: 100px; }
    p { line-height: 20px; }
</style>

<script src="../../resources/check-layout.js"></script>
<body onload="checkLayout('.flexbox')">
<p>This test flex item should ignore container's firstLetter pseudo element.</p>

<div class="container">
    <div class="flexbox">
        <p data-expected-height=20>The first item.</p>
        <p data-expected-height=20>The second item.</p>
    </div>
</div>
</body>
</html>
